requirements.txt
# AI-Driven Unified Data Platform for Oceanographic & Biodiversity Insights
# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data Processing
xarray>=2023.1.0
netCDF4>=1.6.0
rioxarray>=0.14.0
geopandas>=0.13.0

# ML / Clustering
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0

# Config
pyyaml>=6.0

# Scheduling
apscheduler>=3.10.0

# Storage
pyarrow>=12.0.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0

# ── Data Source Clients ──────────────────────────
python-dotenv>=1.0.0          # .env file management
requests>=2.31.0              # HTTP client for ERDDAP/CHIRPS

# NASA Earthdata (optional, for authenticated access)
# Install manually: pip install earthaccess
# earthaccess>=0.8.0

# Copernicus Marine Service (optional, needs credentials)
# Install manually: pip install copernicusmarine
# copernicusmarine>=1.0.0
